Change data count member for BB files to long (=> unsigned int) to accomodate very large files (e.g. Japser TFBS)

+++ src/main/java/org/broad/igv/ucsc/bb/BBHeader.java
@@ -16,5 +16,5 @@ public class BBHeader {
     public long extensionOffset;
     public int extraIndexCount = 0;
     public long[] extraIndexOffsets;
-    public int dataCount;
+    public long dataCount;
 }
